# Improve performance of Vocoder