--- a/src/vocoder/mod.rs
+++ b/src/vocoder/mod.rs
@@ -45,6 +45,7 @@ impl Vocoder {
         }
     }
 
+    #[allow(clippy::too_many_arguments)]
     pub fn synthesize(
         &mut self,
         lf0: f64,
@@ -53,7 +54,8 @@ impl Vocoder {
         alpha: f64,
         beta: f64,
         volume: f64,
-    ) -> Vec<f64> {
+        rawdata: &mut [f64],
+    ) {
         let p = if lf0 == NODATA {
             0.0
         } else if lf0 <= MIN_LF0 {
@@ -107,24 +109,20 @@ impl Vocoder {
                     .get_or_insert_with(|| Excitation::new(p, lpf.len()));
                 excitation.start(p, self.fperiod);
 
-                let rawdata = (0..self.fperiod)
-                    .map(|_| {
-                        let mut x = excitation.get(lpf);
-                        if x != 0.0 {
-                            x *= coefficients[0].exp();
-                        }
-                        filter.df(&mut x, alpha, coefficients);
-                        for i in 0..coefficients.len() {
-                            coefficients[i] += cinc[i];
-                        }
-                        x * volume
-                    })
-                    .collect();
+                (0..self.fperiod).for_each(|i| {
+                    let mut x = excitation.get(lpf);
+                    if x != 0.0 {
+                        x *= coefficients[0].exp();
+                    }
+                    filter.df(&mut x, alpha, coefficients);
+                    for i in 0..coefficients.len() {
+                        coefficients[i] += cinc[i];
+                    }
+                    rawdata[i] = x * volume;
+                });
 
                 excitation.end(p);
                 *coefficients = cc;
-
-                rawdata
             }
             Stage::NonZero {
                 stage,
@@ -151,22 +149,18 @@ impl Vocoder {
                     .get_or_insert_with(|| Excitation::new(p, lpf.len()));
                 excitation.start(p, self.fperiod);
 
-                let rawdata = (0..self.fperiod)
-                    .map(|_| {
-                        let mut x = excitation.get(lpf);
-                        x *= coefficients[0];
-                        filter.df(&mut x, alpha, coefficients);
-                        for i in 0..coefficients.len() {
-                            coefficients[i] += cinc[i];
-                        }
-                        x * volume
-                    })
-                    .collect();
+                (0..self.fperiod).for_each(|i| {
+                    let mut x = excitation.get(lpf);
+                    x *= coefficients[0];
+                    filter.df(&mut x, alpha, coefficients);
+                    for i in 0..coefficients.len() {
+                        coefficients[i] += cinc[i];
+                    }
+                    rawdata[i] = x * volume;
+                });
 
                 excitation.end(p);
                 *coefficients = cc;
-
-                rawdata
             }
         }
     }