main
```
test bonsai        ... bench:  19,259,586 ns/iter (+/- 117,963)
test bonsai_letter ... bench:  53,474,059 ns/iter (+/- 135,598)
test is_bonsai     ... bench:  29,690,178 ns/iter (+/- 103,336)
```
53cebc5
```
test bonsai        ... bench:  18,781,180 ns/iter (+/- 83,165)
test bonsai_letter ... bench:  52,141,404 ns/iter (+/- 88,456)
test is_bonsai     ... bench:  28,981,244 ns/iter (+/- 101,065)
```
4e65ae3
```
test bonsai        ... bench:  17,744,941 ns/iter (+/- 45,752)
test bonsai_letter ... bench:  49,492,881 ns/iter (+/- 135,639)
test is_bonsai     ... bench:  27,443,671 ns/iter (+/- 51,153)
```